Change type of structure XArchive member files_size
Change type of related variables as well.

Additionally, use a proper format string for printing it.

+++ src/archive.h
@@ -131,7 +131,7 @@ struct _XArchive
 	gboolean create_image;
 	/* properties */
 	gint nr_of_files;
-	unsigned long long int files_size;
+	guint64 files_size;
 	gboolean has_passwd;
 	gchar *passwd;
 	gboolean has_comment;
